・特定の環境でダウンロード速度が遅くなることがあるのを修正 #14
・"このPCからの接続のみ許可する"オプションを追加
・コンパイラの警告を修正
・読み込みバッファを増やした
・バッファのコピー回数を減らした
・データの送信中に接続を切ろうとしたり終了しようとしても終了しないバグを修正
・サーバーから送られてくるルートCA証明書がおかしいとき"ASN no signer error to confirm failure"の警告が出るのを修正

+++ Proxydomo/proximodo/headerfilter.cpp
@@ -106,7 +106,7 @@ bool CHeaderFilter::filter(string& content) {
     size_t index = 0;
     while ((index = content.find_first_of("\r\n", index)) != string::npos) {
         // Multiligne must be formatted as CR LF SPACE
-        int pos = index+1;
+        size_t pos = index + 1;
         while (string(" \t\r\n").find(content[pos])) ++pos;
         content.replace(index, pos-index, "\r\n ");
         index += 3;
